feat(admin): 관리 기능을 별도 주소로 빼고 진입에 관리자 자격을 요구한다
관리 권한이 검수자 역할에 통째로 얹혀 있었다. 사용자 역할을 바꾸는 일과 라벨을
검수하는 일이 같은 자격으로 되니, 검수를 맡기려면 계정을 만들고 지우는 권한까지
함께 주어야 했다.

역할을 새로 만들지 않는다. 겸직이 구조적으로 막히기 때문이다 — 역할 저장이 사용자당
한 행이고 그 번호는 상위 시스템이 발급하므로, 겸직자에게 계정을 둘 주려면 우리가
만들 수 없는 것을 상위 시스템에 요청해야 한다. 사용자 관리를 새 역할 전용으로 옮기면
그 역할을 줄 수 있는 사람이 그 역할 보유자뿐이 되는 순환도 생긴다.

대신 관리 기능을 별도 주소로 모으고 진입에 관리자 자격을 요구한다. 자격이 여는 것은
역할 승격이 아니라 그 사람에게 잠깐 열리는 유효창이며, 검수자 권한은 여전히 필요하다.
인가는 대체되지 않고 가산된다. 공유 자격인데도 누가 했는지가 개인 단위로 남는 것은
그 창이 발급 대상자 본인에게 결박돼 요청이 여전히 본인 자격으로 인증되기 때문이다.

요구는 쓰기에만 건다. 조회를 함께 막으면 작업자 목록을 읽는 배정 화면이 깨진다.
이어올리기 업로드도 시작 한 번에만 요구한다 — 조각마다 요구하면 큰 영상이 유효창을
넘겨 구조적으로 올릴 수 없다. 그렇게 좁혀도 안전한 이유는 업로드를 이어 쓰려면
소유자여야 하고 그 소유자가 곧 유효창을 연 사람이기 때문이다.

요구를 선언하는 자리와 판정하는 자리를 나눴다. 어느 요청에 걸리는지는 창구가 표식으로
선언하고, 그 선언을 읽어 판정 지점에 넘기는 것은 공통 배선이 한다. 판정은 한 곳만
소유해 복제되지 않는다. 다만 설정 저장은 요구가 창구가 아니라 설정 키 단위로 갈리므로
표식을 쓰지 않고 판정 지점을 직접 부른다 — 표식은 창구 전체에 걸려서, 붙이면 요구가
없는 키까지 함께 잠긴다.

자격을 운영 중에 바꿀 수 있게 저장 위치를 배포 설정에서 전용 표로 옮겼다. 공용 설정
목록에 두면 그 해시가 검수자 전원에게 보인다. 교체에는 현재 자격 재확인을 함께 요구한다
— 유효창만으로 바꿀 수 있으면 창을 뺏은 것이 곧 자격을 뺏은 것이 된다. 교체하면 그
전에 발급된 창이 전부 무효가 되는데, 세대를 세는 컬럼을 두는 대신 서명키가 현재 자격에
의존하게 해서 얻는다. 상태를 늘리지 않고 같은 결과를 낸다.

되돌림으로 확인했다. 표식을 창구를 담은 타입에 옮기면 조회가 통째로 죽고, 조각 올리기가
잠기며, 본인 자산 업로드까지 끌려 들어간다. 서명키에서 자격을 빼면 무효화가 시늉이 된다.
현재 자격 재확인을 무력화하면 교체가 열린다. 그 각각을 막는 가드가 실재함을 실패로 보였다.

+++ backend/src/main/resources/db/migration/V21__add_ls_mngr_pswd.sql
@@ -18,6 +18,11 @@
 --
 --   수정자아이디는 이 저장소의 기존 mdfr_id 와 같은 폭(30)을 쓴다.
 --
+--   ★타입·크기도 표준도메인을 따른다 — 일련번호는 <bigint> 다 (사업표준 도메인 일련번호B20).
+--   CSV 정본의 일련번호 계열은 일련번호B20(BIGINT) · 일련번호N20/N10/N22(NUMERIC) · 일련번호C2(문자)
+--   뿐이라 smallint 는 어디에도 없고, 이 저장소 베이스라인의 _SN 컬럼 102개도 전부 bigint 다.
+--   「행이 1개뿐이니 좁혀도 된다」는 판단은 표준도메인을 임의로 바꾸는 것이라 감리 지적 대상이다.
+--
 -- ----------------------------------------------------------------------------
 -- * ★ 단일 행 강제 — PK + CHECK 두 겹이 <함께> 있어야 성립한다
 -- ----------------------------------------------------------------------------
@@ -81,7 +86,7 @@
 -- ============================================================================
 
 CREATE TABLE IF NOT EXISTS ls_mngr_pswd (
-    mngr_pswd_sn smallint NOT NULL,
+    mngr_pswd_sn bigint NOT NULL,
     pswd_hash character varying(100) NOT NULL,
     mdfr_id character varying(30),
     pswd_mdfcn_dt timestamp without time zone
